profile

+++ components/LikeBtn.jsx
@@ -1,4 +1,4 @@
-import { collection, deleteDoc, doc, getDocs, onSnapshot, query, setDoc } from 'firebase/firestore';
+import { collection, deleteDoc, doc, getDocs, onSnapshot, query, serverTimestamp, setDoc } from 'firebase/firestore';
 import { useSession } from 'next-auth/react';
 import React, { useEffect, useState } from 'react'
 import { db } from '../firebase';
@@ -8,6 +8,7 @@ function LikeBtn({ id }) {
 	const { data: session } = useSession();
 	const [likes, setLikes] = useState([]);
 	const [hasLike, setHasLike] = useState(false);
+	const [videos, setVideos] = useState([]);
 
   useEffect(() => {
 		if (id) {
@@ -18,6 +19,8 @@ function LikeBtn({ id }) {
 						setLikes(snapshot.docs);
 					}
 			);
+
+			
 			
 			return likeUnSub;
 
@@ -25,7 +28,23 @@ function LikeBtn({ id }) {
     
    
     
-	}, [db,id]);
+	}, [db, id]);
+
+
+	useEffect(() => { 
+const q = query(collection(db, "tiktot_posts"));
+
+const unVideosubscribe = onSnapshot(q, (querySnapshot) => {
+	const video = [];
+	querySnapshot.forEach((doc) => {
+		video.push({ ...doc.data(), videoId: doc.id });
+		setVideos(video);
+	});
+});
+		return unVideosubscribe;
+	}, [db]);
+	
+	console.log(videos);
 
 	useEffect(() => {
 		
@@ -36,11 +55,13 @@ function LikeBtn({ id }) {
   }, [likes]);
   
   const likeVideo = async () => {
-		if (session) {
+		if (session && videos) {
 			if (hasLike) {
 				await deleteDoc(
 					doc(db, "tiktot_posts", id, "like", session?.user?.uid)
 				);
+
+				await deleteDoc(doc(db, "tiktot_likes", session?.user?.uid));
 			} else {
 				await setDoc(
 					doc(db, "tiktot_posts", id, "like", session?.user?.uid),
@@ -48,6 +69,13 @@ function LikeBtn({ id }) {
 						username: session.user?.username,
 					}
 				);
+
+				await setDoc(doc(db, "tiktot_likes", session?.user?.uid), {
+					videoId: videos.videoId || null,
+					video: videos.video || null,
+					caption: videos.caption || null,
+					timestamp: serverTimestamp(),
+				});
 			}
 		}
 	};
